requirements.txt
aiofiles==24.1.0
aiohttp==3.11.11
regex==2023.12.25
typing_extensions==4.13.2
anyio==4.8.0
pydantic==2.11.5
python-dotenv==1.1.0
beautifulsoup4==4.13.4
charset_normalizer==3.4.2
trafilatura==2.0.0
openai==1.82.1
mcp==1.6.0
pandas==2.2.3
PyPDF2==3.0.1
python_docx==1.1.2
python_pptx==1.0.0
playwright==1.49.1
numpy==1.26.4
sympy==1.13.1
pandas==2.2.3
matplotlib==3.8.4
plotly==5.24.1
seaborn==0.13.2
scikit-learn==1.4.2
statsmodels==0.14.4
prophet==1.1.6
chainlit==2.5.5